+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Next
 ## New features
 
 - New option --print-srcaddr to show used IP source address in output (#461, thanks @gsnw-sebast)
+- Add IPv6 support to --print-srcaddr (#462, thanks @gsnw-sebast)
 
 ## Bugfixes and other changes